code/src/EmailClassification/requirements.txt
pdfplumber
pytesseract
python-docx
pillow
pymupdf
docx2txt
faiss-cpu
sentence-transformers
langchain
bitsandbytes
accelerate
transformers
huggingface_hub
numpy
